Feature/#24 reservation biz logic

+++ src/main/java/ybe/mini/travelserver/domain/reservation_room/dto/ReservationRoomCreateRequest.java
@@ -1,12 +1,6 @@
 package ybe.mini.travelserver.domain.reservation_room.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
-import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
 
 import java.io.Serializable;
 import java.time.LocalDateTime;
@@ -18,6 +12,6 @@ public record ReservationRoomCreateRequest(
 
         @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm")
         LocalDateTime checkOut,
-        Long guestNumber
+        Integer guestNumber
 ) implements Serializable {
 }
